Revert "添加对asp.net的支持 (#3)"
This reverts commit 667c3292eda0bc3198698da9ff07786238148a61.
