+++ ICanPay.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.27004.2010
+VisualStudioVersion = 15.0.27004.2009
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ICanPay.Alipay", "src\ICanPay.Alipay\ICanPay.Alipay.csproj", "{C1A4A406-8346-4BA8-A587-91F305AD3668}"
 EndProject
@@ -14,10 +14,6 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ICanPay.UnitTest", "test\ICanPay.UnitTest\ICanPay.UnitTest.csproj", "{639AA7FB-AAB4-44A5-84AF-F28D9ED59573}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ICanPay.Unionpay", "src\ICanPay.Unionpay\ICanPay.Unionpay.csproj", "{D0CBE889-BEAE-4F88-BF23-A4F1ED728620}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ICanPay.UnitTest(.Net)", "test\ICanPay.UnitTest(.Net)\ICanPay.UnitTest(.Net).csproj", "{28AA717D-D840-4C92-A2AD-AFBF3794D6EF}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ICanPay.Demo(Net)", "sample\ICanPay.Demo(Net)\ICanPay.Demo(Net).csproj", "{65D7A296-D254-4B30-82C8-CF5C1B4C8D7B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -49,14 +45,6 @@ Global
 		{D0CBE889-BEAE-4F88-BF23-A4F1ED728620}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D0CBE889-BEAE-4F88-BF23-A4F1ED728620}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D0CBE889-BEAE-4F88-BF23-A4F1ED728620}.Release|Any CPU.Build.0 = Release|Any CPU
-		{28AA717D-D840-4C92-A2AD-AFBF3794D6EF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{28AA717D-D840-4C92-A2AD-AFBF3794D6EF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{28AA717D-D840-4C92-A2AD-AFBF3794D6EF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{28AA717D-D840-4C92-A2AD-AFBF3794D6EF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{65D7A296-D254-4B30-82C8-CF5C1B4C8D7B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{65D7A296-D254-4B30-82C8-CF5C1B4C8D7B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{65D7A296-D254-4B30-82C8-CF5C1B4C8D7B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{65D7A296-D254-4B30-82C8-CF5C1B4C8D7B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
