--- sample/ICanPay.Demo(Net)/Global.asax
@@ -1,1 +1,0 @@
-﻿<%@ Application Codebehind="Global.asax.cs" Inherits="ICanPay.Demo_Net_.Global" Language="C#" %>
