+++ src/ICanPay.Core/Gateways/GatewayData.cs
@@ -1,12 +1,9 @@
-﻿#if NETSTANDARD2_0
+﻿using ICanPay.Core.Utils;
 using Microsoft.AspNetCore.Http;
-#endif
-using ICanPay.Core.Utils;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
-using System.Collections.Specialized;
 using System.Linq;
 using System.Net;
 using System.Reflection;
@@ -22,14 +19,14 @@ namespace ICanPay.Core
     /// </summary>
     public class GatewayData
     {
-#region 私有字段
+        #region 私有字段
 
         private readonly SortedDictionary<string, object> _values;
         private string _originalResult = null;
 
-#endregion
+        #endregion
 
-#region 属性
+        #region 属性
 
         public object this[string key]
         {
@@ -39,9 +36,9 @@ namespace ICanPay.Core
 
         public int Count => _values.Count;
 
-#endregion
+        #endregion
 
-#region 构造函数
+        #region 构造函数
 
         /// <summary>
         /// 构造函数
@@ -60,9 +57,9 @@ namespace ICanPay.Core
             _values = new SortedDictionary<string, object>(comparer);
         }
 
-#endregion
+        #endregion
 
-#region 方法
+        #region 方法
 
         /// <summary>
         /// 添加参数
@@ -403,8 +400,6 @@ namespace ICanPay.Core
             }
         }
 
-#if NETSTANDARD2_0
-
         /// <summary>
         /// 将表单数据转换为网关数据
         /// </summary>
@@ -423,20 +418,6 @@ namespace ICanPay.Core
                 }
             }
             catch { }
-        }
-#endif
-
-        /// <summary>
-        /// 将键值对转换为网关数据
-        /// </summary>
-        /// <param name="nvc">键值对</param>
-        /// <returns></returns>
-        public void FromNameValueCollection(NameValueCollection nvc)
-        {
-            foreach (var item in nvc.AllKeys)
-            {
-                Add(item, nvc[item]);
-            }
         }
 
         /// <summary>
@@ -580,6 +561,6 @@ namespace ICanPay.Core
         /// <returns></returns>
         public string GetOriginalResult() => _originalResult;
 
-#endregion
+        #endregion
     }
 }
