+++ src/ICanPay.Core/Payments/NotifyProcess.cs
@@ -92,11 +92,7 @@ namespace ICanPay.Core
                 {
                     try
                     {
-#if NETSTANDARD2_0
                         gatewayData.FromForm(HttpUtil.Form);
-#else
-                        gatewayData.FromNameValueCollection(HttpUtil.Form);
-#endif
                     }
                     catch { }
                 }
