+++ src/ICanPay.Wechatpay/ICanPay.Wechatpay.csproj
@@ -1,13 +1,13 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net45;netstandard2.0</TargetFrameworks>
+    <TargetFramework>netstandard2.0</TargetFramework>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <PackageProjectUrl>https://github.com/Varorbc/ICanPay</PackageProjectUrl>
     <Title>ICanPay.Wechatpay</Title>
     <Authors>Varorbc</Authors>
     <Description>微信支付</Description>
-    <Version>1.1.5</Version>
+    <Version>1.1.4</Version>
     <PackageTags>dotnetcore;pay;wechatpay;</PackageTags>
     <RepositoryType>git</RepositoryType>
     <RepositoryUrl>https://github.com/Varorbc/ICanPay</RepositoryUrl>
